requirements.txt
python-slugify==3.0.2
hdx-python-api==3.7.4
pandas==0.23.4
-r docker-requirements.txt